* commit '134717b87ba5f1b9e00f2f27e9cfbbd5590b7012':
  VFS-9300 Resolving PR tasks
  VFS-9300 checksum-counting-oneclient workflow test metasteps
  VFS-9300 fix test: checksum-counting-oneclient workflow
  VFS-9300 Scenario completed
  VFS-8849 Adding parallel box below/above
  VFS-9300 Changes in scenario
  VFS-8849 Scenarios fixes
  VFS-8849 Scenario

+++ test_run.py
@@ -366,7 +366,8 @@ ALL       ALL = (ALL) NOPASSWD: ALL
     if env_errors_exists(report):
         ret = 1
 
-    shutil.rmtree(tmp_user_config_dir)
+    if 'tmp_user_config_dir' in locals():
+        shutil.rmtree(tmp_user_config_dir)
 
     sys.exit(ret)
